requirements.txt
google-generativeai==0.8.3
PyPDF2==3.0.1
streamlit==1.41.1
python-dotenv==1.0.0